moved fit methods to a parent class
